+++ R/globals.R
@@ -1,0 +1,4 @@
+globalVariables(c(
+  "meta_levels", "name", "name_in_survey", "question_name", "question_type",
+  "selector_supp", "supported", "type_supp", "value"
+))