+++ NEWS.md
@@ -1,3 +1,7 @@
+# qualtRics 3.1.1 [WIP]
+
+- 
+
 # qualtRics 3.1.0
 
 - New maintainer: Julia Silge